+++ lib/components/builders/conditions/if.dart
@@ -1,3 +1,0 @@
-import 'package:pipen/components/builders/conditions/pipen_build_if.dart';
-
-typedef If = PipenBuildIf;