+++ lib/extensions/list_extension.dart
@@ -1,7 +1,0 @@
-extension ListExtension on List {
-  List<T> cloneRemove<T>(Object? value) {
-    List<T> clone = [...this as List<T>];
-    clone.remove(value);
-    return clone;
-  }
-}