+++ lib/listener/listener_state.dart
@@ -1,3 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-typedef ListenerState<T> = (BuildContext context, T state);