+++ lib/src/components/gesture/pipen_gesture_dismiss_keyboard.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/widgets.dart';
+
+class PipenGestureDismissKeyboard extends StatelessWidget {
+  const PipenGestureDismissKeyboard({required this.child, super.key});
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) => GestureDetector(
+    onTap: () {
+      FocusScopeNode currentFocus = FocusScope.of(context);
+      if (!currentFocus.hasPrimaryFocus && currentFocus.focusedChild != null) {
+        FocusManager.instance.primaryFocus?.unfocus();
+      }
+    },
+    child: child,
+  );
+}